% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainclass.R
\name{brainclass}
\alias{brainclass}
\title{Transcriptional data guided fMRI network classification}
\usage{
brainclass(
  X,
  y,
  edgeGrp,
  family = "binomial",
  penalty = "grLasso",
  lambda = 0.005
)
}
\arguments{
\item{X}{A matrix with the training samples, in wich each row represents the
vectorized (by column order) upper triangular part of a network adjacency
matrix.}

\item{y}{A vector containing the class labels of the training samples (2
classes by default).}

\item{edgeGrp}{A list of vectors, each containing integer indices or
character names of variables in the group.}

\item{family}{See grpregOverlap::grpregOverlap for details.
Default is "binomial".}

\item{penalty}{See grpregOverlap::grpregOverlap for details.
Default is "grLasso".}

\item{lambda}{See grpregOverlap::grpregOverlap for details.
Default is 5e-3.}
}
\value{
An object with S3 class "grpregOverlap".
See grpregOverlap::`grpregOverlap` for details.
}
\description{
The `brainclass` function performs fMRI network classification where edges
in each fMRI network are treated as a long vector and used as features.
}
\examples{
# Load a toy data set.
data(toyData)
res <- brainclass(X = toyData$X, y = toyData$y, edgeGrp = toyData$grpList)

}
\references{
Zeng,Y. and Breheny,P. (2016) Overlapping group logistic regression with
applications to genetic pathway selection. Cancer Informatics, 15,
CIN-S40043.
}
