% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneExpr.R
\name{filterGeneExpr}
\alias{filterGeneExpr}
\title{Filter gene expression data by average expression}
\usage{
filterGeneExpr(geneExpr, q = 0.75)
}
\arguments{
\item{geneExpr}{The ROI-by-gene data frame or matrix.}

\item{q}{The quantile cutoff.}
}
\value{
Downsized gene expression data.
}
\description{
The `filterGeneExpr` function applies downsizing on the gene expression data
by average expression.
}
\examples{
## NOT RUN
# filterGeneExpr(ahba, q = 0.75)

}
