% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneSets.R
\name{filterGeneSets}
\alias{filterGeneSets}
\title{Filter gene sets by size and presence of each gene in the reference gene
expression data set}
\usage{
filterGeneSets(geneSetList, candidateGenes, min.size = 25, max.size = 500)
}
\arguments{
\item{geneSetList}{A gene set collection stored in a list.}

\item{candidateGenes}{A character vector of all genes to be considered.}

\item{min.size}{Minimum number of genes to keep a gene set. Cannot be smaller
than 2.}

\item{max.size}{Maximum number of genes to keep a gene set.}
}
\value{
Filtered list of gene sets. This output can be used to construct gene
set edge groups with the `getGeneSetEdgeGroup` function.
}
\description{
The `filterGeneSets` function filters gene sets by their sizes, and filters
out genes not in the candidate list.
}
\examples{
## NOT RUN
# filterGeneSets(gscv7.0$kegg, colnames(ahba), min.size = 5, max.size = Inf)

}
