% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVfolds.R
\name{getCVfolds}
\alias{getCVfolds}
\title{Get indices of the observations assigned to each fold for repeated k-fold
cross-validation}
\usage{
getCVfolds(y, k = 10, repeats = 50, type = "random")
}
\arguments{
\item{y}{A binary vector of length n.}

\item{k}{An integer giving the number of groups into which n observations
need to be split. The default is 10.}

\item{repeats}{Number of repeats. Default is 50.}

\item{type}{A character string to specify the type of folds to be obtained.
See cvTools::cvFolds for more. The default is "random".}
}
\value{
A list object of data frames, each consisting of two columns, where
column k gives the fold for each premuted observation; column Ind gives the
index of the observation. Each data frame element of the returned list can be
used in a different repeat of k-fold cross-validation.
}
\description{
The `getCVfolds` function splits n observations into k groups in s different
ways to be used for repeated k-fold cross-validations over s repeats. This
function is designed for situations where the the response variable is
binary.
}
\examples{
## NOT RUN
# getCVfolds(toyData$y, k = 3, repeats = 5, type = "random")

}
