% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneSetEdgeGroup.R
\name{getGeneSetEdgeGroup}
\alias{getGeneSetEdgeGroup}
\title{Construct gene set edge groups for a gene set collection (GSC) of interest}
\usage{
getGeneSetEdgeGroup(geneExpr, geneSetList, cutoff = 0.99)
}
\arguments{
\item{geneExpr}{The ROI-by-gene data frame or matrix. Potentially the output
from the `filterGeneExpr` function.}

\item{geneSetList}{A gene set collection stored in a list. Potentially the
output from the `filterGeneSets` function}

\item{cutoff}{The hard-thresholding cutoff on edge weights.}
}
\value{
A list object of vectors, each containing labels of edges that are
the most highly weighted in the corresponding gene set expression network.
Each item in the returned list represents a gene set edge group. These gene
set edge groups are used as feature groups in `brainclass` classification.
}
\description{
The `getGeneSetEdgeGroup` function generates the gene set edge groups in
`brainclass` classification.
}
\examples{
## NOT RUN
# data(ahba)
# data(gscv7.0)
# ahba <- filterGeneExpr(ahba)
# kegg <- filterGeneSets(geneSetList = gscv7.0$kegg,
# candidateGenes = colnames(ahba), min.size = 5, max.size = Inf)
# keggEdgeGrp <- getGeneSetEdgeGroup(geneExpr = ahba, geneSetList = kegg)

}
