% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelectedGroup.R
\name{getSelectedGroup}
\alias{getSelectedGroup}
\title{Identify feature groups where all fitted coefficients are non-zero.}
\usage{
getSelectedGroup(fit)
}
\arguments{
\item{fit}{The `brainclass` fit.}
}
\value{
A character vector. Names of edge groups where all fitted
coefficients are non-zero.
}
\description{
The `getSelectedGroup` function extracts names of the features groups where
all edges have non-zero fitted coefficients.
}
\note{
For the data example, see \link[brainClass]{brainclass}.
}
