% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gscv7.0}
\alias{gscv7.0}
\title{Gene set collections downloaded from MSigDB v7.0}
\format{
A list object.
\describe{
\item{gscv7.0}{A list object of gene set collections of different types.
Here we included 8 different collections, including the positional (by
gene location on each chromosome). KEGG pathways, REACTOME pathways,
transcription factor targets, immunological signatures and three classes of
geno ontology terms (biological processes, cellular components and molecular
functions). }
}
}
\usage{
data("gscv7.0", package = "brainClass")
}
\description{
Gene set collections in entrez ID. Downloaded from the MSigDB database
version 7.0
}
\examples{
data("gscv7.0", package = "brainClass")
}
\references{
Liberzon, A., Subramanian, A., Pinchback, R., Thorvaldsdóttir, H., Tamayo,
P., & Mesirov, J. P. (2011). Molecular signatures database (MSigDB) 3.0.
*Bioinformatics*, 27(12), 1739-1740.
}
\keyword{datasets}
