% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\title{Calculate Jaccard Index of two sets}
\usage{
jaccard(x, y)
}
\arguments{
\item{x, y}{Vectors of the same node containing a sequence of non-duplicated
items.}
}
\value{
The Jaccard Index as a numeric.
}
\description{
The `jaccard` function calculates the Jaccard Index of two sets.
}
\examples{
## NOT RUN
# jaccard(1:10, 1:20)

}
