% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{A toy data set for the `brainclass` function}
\format{
A list object with three elements
\describe{
\item{X}{A matrix containing the upper triangular part of the fMRI networks
of subjects. Each row represents a subject and columns represent edges. }
\item{y}{Class labels of the subjects in the dataset. \code{y = 1} represents
schizophrenic. }
\item{grpList}{Example gene set edge groups stored in a list, whose elements
are character vectors of edge labels belonging to each feature group. }
\item{toyGenExpr}{A small submatrix from the (pre-processed) Allen Human
Brain Atlas (AHBA) brain transcriptional data. }
\item{exampleGSC}{A example of the format of a gene set collection in the
format of a list. This list should contain the gene names or IDs consisting
of a gene set from the selected gene set collection. Also the type of gene
ID should match with the format of gene names in the column of the
transcriptional data matrix. }
}
}
\usage{
data(toyData, package = "brainClass")
}
\description{
A subset of the COBRE data set to be used for examples on how to use the
`brainclass` function
}
\examples{
data(toyData, package = "brainClass")
}
\keyword{datasets}
