% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahba}
\alias{ahba}
\title{Processed Allen Human Brain Atlas (AHBA) data}
\format{
A data frame object.
\describe{
\item{ahba}{A data frame containing the gene expression information of ROIS
defined by the Power parcellation. The data frame is organised with gene
(entrez IDs) on the columns and ROI on the rows. }
}
}
\usage{
data("ahba", package = "brainClass")
}
\description{
Processed whole brain gene expression data. AHBA samples are mapped into
regions of interest (ROIs) defined by Power parcellation.
}
\examples{
data("ahba", package = "brainClass")
}
\references{
Hawrylycz, M. J., Lein, E. S., Guillozet-Bongaarts, A. L., Shen, E. H., Ng,
L., Miller, J. A., ... & Jones, A. R. (2012). An anatomically comprehensive
atlas of the adult human brain transcriptome. \emph{Nature}, 489(7416), 391-399.

Power, J. D., Cohen, A. L., Nelson, S. M., Wig, G. S., Barnes, K. A., Church,
J. A., ... & Petersen, S. E. (2011). Functional network organization of the
human brain. \emph{Neuron}, 72(4), 665-678.
}
\keyword{datasets}
