% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digsee}
\alias{digsee}
\title{Processed DiGSeE scores of relevance to schizophrenia of each gene set
signature}
\format{
A list object.
\describe{
\item{digsee}{A list object of data frames with summarised DiGSeE
schizophrenia relevance scores for different gene set collections.
Each data frame contains three columns. `prop` means the proportion of
genes with an effective DiGSeE score. `sumScore` is the sum of all scores;
and `avgScore` stores the mean score of each gene set. }
}
}
\usage{
data("digsee", package = "brainClass")
}
\description{
Summarised schizophrenia relevance score of each gene set from the genewise
scores downloaded from DiGSeE database. Scores were log-transformed. Details
on the processing of raw scores are available in the Supplementary Methods in
the reference.
}
\examples{
data("digsee", package = "brainClass")
}
\references{
Kim, J., So, S., Lee, H. J., Park, J. C., Kim, J. J., & Lee, H. (2013).
DigSee: disease gene search engine with evidence sentences (version cancer).
\emph{Nucleic acids research}, 41(W1), W510-W517.

Li, M., Kessler, D., Arroyo, J., Freytag, S., Bahlo, M., Levina, E., & Yang,
J. Y. H. (2020). Guiding and interpreting brain network classification with
transcriptional data. \emph{bioRxiv}.
}
\keyword{datasets}
