% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneExpr.R
\name{filterGeneExpr}
\alias{filterGeneExpr}
\title{Filter gene expression data by average expression in each region of interest
 (ROI)}
\usage{
filterGeneExpr(geneExpr, q = 0.75)
}
\arguments{
\item{geneExpr}{The ROI-by-gene data frame.}

\item{q}{The quantile cutoff.}
}
\value{
Downsized gene expression data frame.
}
\description{
The `filterGeneExpr` function
}
\examples{
## NOT RUN
# filterGeneExpr(ahba, q = 0.75)

}
