% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeLabel.R
\name{getEdgeLabel}
\alias{getEdgeLabel}
\title{Obtain labels of edges in a network}
\usage{
getEdgeLabel(node, sep = "-")
}
\arguments{
\item{node}{A numeric or character vector, which contains the names of
all nodes in a network.}

\item{sep}{The seperation pattern. Default is "-".}
}
\value{
A character vector of all edge labels for each network.
}
\description{
The `getEdgeLabel` function generates the labels of edges in a network upon
names of the nodes in the network.
}
\examples{
getEdgeLabel(c(1:10, 12:20))
getEdgeLabel(paste("ROI", 1:5, sep = ""))
# In COBRE data
getEdgeLabel(c(1:74, 76:264))

}
