% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROC.R
\name{getROC}
\alias{getROC}
\title{Obtain receiver operating characteristic (ROC) curve}
\usage{
getROC(truth, test.metric, step.size = 0.05)
}
\arguments{
\item{metrics}{The data frame storing}
}
\value{

}
\description{
The `getROC` function calculates
}
\examples{
## NOT RUN
# jaccard(1:10, 1:20)

}
