% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.edge.R
\name{posthoc.edge}
\alias{posthoc.edge}
\title{Obtain post-hoc interpretability metrics by edges}
\usage{
posthoc.edge(
  selected.edgeLabels,
  all.edgeLabels,
  geneSetList,
  geneExpr,
  get.jaccard = TRUE,
  jaccard.cutoff = 0.99,
  get.betweenness = TRUE,
  betweenness.cutoff = 0.75,
  iter = 200,
  adjust.p = TRUE,
  ...
)
}
\arguments{
\item{selected.edgeLabels}{}

\item{all.edgeLabels}{}

\item{geneSetList}{}

\item{geneExpr}{}

\item{jaccard.cutoff}{}

\item{betweenness.cutoff}{}

\item{iter}{}

\item{adjust.p}{Logical. Whether or not to return adjusted p-values.}

\item{...}{}
}
\value{
Evalutaion metrics stored in a data frame with four columns.
}
\description{
The `posthoc.edge` function calculates the edgewise interpreration metrics on
set of edges selected by the generic network classifier.
}
\examples{
## NOT RUN
#

}
\references{
Li, M., Kessler, D., Arroyo, J., Freytag, S., Bahlo, M., Levina, E., & Yang,
J. Y. H. (2020). Guiding and interpreting brain network classification with
transcriptional data. *bioRxiv*.

Csardi, G., & Nepusz, T. (2006). The igraph software package for complex
network research. *InterJournal, complex systems*, 1695(5), 1-9.
}
